+++ doc/zephyr.doxyfile.in
@@ -759,7 +759,6 @@ INPUT                  = @ZEPHYR_BASE@/include/ \
                          @ZEPHYR_BASE@/include/arch/arm/cortex_m \
                          @ZEPHYR_BASE@/include/arch/nios2/ \
                          @ZEPHYR_BASE@/lib/libc/minimal/include/ \
-                         @ZEPHYR_BASE@/include/net/coap_sock.h \
                          @ZEPHYR_BASE@/include/net/dns_resolve.h \
 			 @ZEPHYR_BASE@/tests/ztest/include/ \
 			 @ZEPHYR_BASE@/tests/kernel/
@@ -808,8 +807,6 @@ EXCLUDE                = @ZEPHYR_BASE@/include/spi_legacy.h \
                          @ZEPHYR_BASE@/include/net/http_legacy.h \
                          @ZEPHYR_BASE@/include/cmsis_rtos_v1/cmsis_os.h \
                          @ZEPHYR_BASE@/include/cmsis_rtos_v2/cmsis_os2.h \
-                         @ZEPHYR_BASE@/include/net/coap.h \
-                         @ZEPHYR_BASE@/include/net/coap_link_format.h \
 
 # The EXCLUDE_SYMLINKS tag can be used to select whether or not files or
 # directories that are symbolic links (a Unix file system feature) are excluded
