+++ include/net/coap_link_format.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 Intel Corporation
+ * Copyright (c) 2018 Intel Corporation
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -29,17 +29,17 @@ extern "C" {
 #define COAP_WELL_KNOWN_CORE_PATH \
 	((const char * const[]) { ".well-known", "core", NULL })
 
-__deprecated int coap_well_known_core_get(struct coap_resource *resource,
-					  struct coap_packet *request,
-					  struct coap_packet *response,
-					  struct net_pkt *pkt);
+int coap_well_known_core_get(struct coap_resource *resource,
+			     struct coap_packet *request,
+			     struct coap_packet *response,
+			     u8_t *data, u16_t len);
 
 /**
  * In case you want to add attributes to the resources included in the
  * 'well-known/core' "virtual" resource, the 'user_data' field should point
  * to a valid coap_core_metadata structure.
  */
-__deprecated struct coap_core_metadata {
+struct coap_core_metadata {
 	const char * const *attributes;
 	void *user_data;
 };
