--- samples/net/coap_client/prj.conf
@@ -1,18 +1,0 @@
-CONFIG_NETWORKING=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_LOG=y
-CONFIG_NET_UDP=y
-CONFIG_ENTROPY_GENERATOR=y
-CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_COAP=y
-CONFIG_COAP_NET_PKT=y
-
-CONFIG_NET_CONFIG_SETTINGS=y
-CONFIG_NET_CONFIG_NEED_IPV6=y
-CONFIG_NET_CONFIG_MY_IPV6_ADDR="2001:db8::2"
-CONFIG_NET_CONFIG_PEER_IPV6_ADDR="2001:db8::1"
-
-CONFIG_NET_LOG=y
-CONFIG_LOG=y
-CONFIG_LOG_RUNTIME_FILTERING=y
-CONFIG_LOG_BUFFER_SIZE=2048
