--- samples/net/coap_server/prj_bt.conf
@@ -1,18 +1,0 @@
-CONFIG_BT=y
-CONFIG_BT_DEBUG_LOG=y
-CONFIG_BT_DEBUG_L2CAP=y
-CONFIG_BT_SMP=y
-CONFIG_BT_SIGNING=y
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_L2CAP_DYNAMIC_CHANNEL=y
-CONFIG_NETWORKING=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_LOG=y
-CONFIG_LOG=y
-CONFIG_NET_UDP=y
-CONFIG_NET_L2_BT=y
-CONFIG_NET_CONFIG_BT_NODE=y
-CONFIG_ENTROPY_GENERATOR=y
-CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_COAP=y
-CONFIG_COAP_NET_PKT=y
