--- samples/net/coap_server/prj_cc2520.conf
@@ -1,13 +1,0 @@
-CONFIG_GPIO=y
-CONFIG_NETWORKING=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_LOG=y
-CONFIG_NET_UDP=y
-CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_IEEE802154_CC2520=y
-CONFIG_COAP=y
-CONFIG_COAP_NET_PKT=y
-CONFIG_NET_CONFIG_SETTINGS=y
-CONFIG_NET_CONFIG_IEEE802154_CHANNEL=20
-CONFIG_NET_CONFIG_MY_IPV6_ADDR="2001:db8::1"
-CONFIG_NET_CONFIG_PEER_IPV6_ADDR="2001:db8::2"
