--- samples/net/coap_server/sample.yaml
@@ -1,18 +1,0 @@
-sample:
-  description: TBD
-  name: TBD
-tests:
-  test:
-    arch_whitelist: x86
-    harness: net
-    tags: net
-  test_bt:
-    harness: net
-    extra_args: CONF_FILE="prj_bt.conf"
-    platform_whitelist: qemu_x86
-    tags: net bluetooth
-  test_net:
-    harness: net
-    extra_args: CONF_FILE="prj_cc2520.conf"
-    platform_whitelist: quark_se_c1000_devboard
-    tags: net
