--- samples/net/leds_demo/prj_802154.conf
@@ -1,12 +1,0 @@
-CONFIG_GPIO=y
-CONFIG_NETWORKING=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_L2_IEEE802154=y
-CONFIG_NET_LOG=y
-CONFIG_NET_UDP=y
-CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_IEEE802154_CC2520=y
-CONFIG_COAP=y
-CONFIG_COAP_NET_PKT=y
-CONFIG_NET_CONFIG_SETTINGS=y
-CONFIG_NET_CONFIG_IEEE802154_CHANNEL=20
