+++ samples/net/sockets/coap_client/prj.conf
@@ -9,8 +9,6 @@ CONFIG_NET_SOCKETS_POSIX_NAMES=y
 CONFIG_NET_SOCKETS_POLL_MAX=4
 
 CONFIG_COAP=y
-# CoAP Over Sockets
-CONFIG_COAP_SOCK=y
 
 # Kernel options
 CONFIG_ENTROPY_GENERATOR=y
