+++ samples/net/sockets/coap_server/prj.conf
@@ -8,9 +8,8 @@ CONFIG_NET_SOCKETS=y
 CONFIG_NET_SOCKETS_POSIX_NAMES=y
 CONFIG_NET_SOCKETS_POLL_MAX=4
 
-# CoAP Over Sockets
+# CoAP
 CONFIG_COAP=y
-CONFIG_COAP_SOCK=y
 CONFIG_COAP_WELL_KNOWN_BLOCK_WISE=n
 
 # Kernel options
