+++ subsys/net/lib/coap/CMakeLists.txt
@@ -1,10 +1,6 @@
 zephyr_include_directories(.)
 
-zephyr_sources_ifdef(CONFIG_COAP_NET_PKT
+zephyr_sources_ifdef(CONFIG_COAP
   coap.c
   coap_link_format.c
-)
-zephyr_sources_ifdef(CONFIG_COAP_SOCK
-  coap_sock.c
-  coap_link_format_sock.c
 )
