+++ tests/net/lib/coap/prj.conf
@@ -1,14 +1,30 @@
+#Testing
 CONFIG_TEST=y
 CONFIG_NET_TEST=y
+
+# Generic networking options
 CONFIG_NETWORKING=y
+CONFIG_NEWLIB_LIBC=y
 CONFIG_NET_IPV6=y
 CONFIG_NET_UDP=y
-CONFIG_NET_TCP=n
-CONFIG_NET_IPV4=n
-CONFIG_NET_BUF_LOG=y
+
+CONFIG_NET_SOCKETS=y
+CONFIG_NET_SOCKETS_POSIX_NAMES=y
+CONFIG_NET_SOCKETS_POLL_MAX=4
+
+# CoAP Over Sockets
+CONFIG_COAP=y
+CONFIG_COAP_WELL_KNOWN_BLOCK_WISE=n
+CONFIG_COAP_TEST_API_ENABLE=y
+
+# Kernel options
 CONFIG_ENTROPY_GENERATOR=y
 CONFIG_TEST_RANDOM_GENERATOR=y
-CONFIG_COAP=y
-CONFIG_COAP_NET_PKT=y
-CONFIG_COAP_TEST_API_ENABLE=y
+
+# Logging
+CONFIG_PRINTK=y
+CONFIG_NET_LOG=y
+CONFIG_COAP_LOG_LEVEL_DBG=y
+
 CONFIG_MAIN_STACK_SIZE=2048
+CONFIG_HEAP_MEM_POOL_SIZE=2048
